[UX] Make the form more accessible
The `<label for="">` attribute requires the input field to have an ID,
not a name (this makes much sense in case of radio buttons, where all
buttons in a group have the same name, but they have different labels).
If the form has an `<input type="button">` instead of a `<button>`, it
can be submitted by pressing the return key on the keyboard, without
tabbing through the inputs or using the mouse.